+++ source/renderer/app/components/status/DaedalusDiagnostics.js
@@ -146,21 +146,6 @@ const messages = defineMessages({
     defaultMessage: '!!!Sync percentage',
     description: 'Sync percentage',
   },
-  remainingUnsyncedBlocks: {
-    id: 'daedalus.diagnostics.dialog.remainingUnsyncedBlocks',
-    defaultMessage: '!!!Remaining unsynced blocks',
-    description: 'Remaining unsynced blocks',
-  },
-  latestLocalBlockAge: {
-    id: 'daedalus.diagnostics.dialog.latestLocalBlockAge',
-    defaultMessage: '!!!Latest local block age',
-    description: 'Latest local block age',
-  },
-  latestNetworkBlockAge: {
-    id: 'daedalus.diagnostics.dialog.latestNetworkBlockAge',
-    defaultMessage: '!!!Latest network block age',
-    description: 'Latest network block age',
-  },
   localTimeDifference: {
     id: 'daedalus.diagnostics.dialog.localTimeDifference',
     defaultMessage: '!!!Local time difference',
@@ -195,16 +180,6 @@ const messages = defineMessages({
     id: 'daedalus.diagnostics.dialog.cardanoNodeStatusRestart',
     defaultMessage: '!!!Restart Cardano node',
     description: 'Restart Cardano node',
-  },
-  cardanoNodeDiagnostics: {
-    id: 'daedalus.diagnostics.dialog.cardanoNodeDiagnostics',
-    defaultMessage: '!!!Cardano node diagnostics',
-    description: 'Cardano node diagnostics',
-  },
-  realtimeStatisticsMonitor: {
-    id: 'daedalus.diagnostics.dialog.realtimeStatisticsMonitor',
-    defaultMessage: '!!!Realtime statistics monitor',
-    description: 'Realtime statistics monitor',
   },
   cardanoNodeState: {
     id: 'daedalus.diagnostics.dialog.cardanoNodeState',
@@ -347,10 +322,6 @@ type Props = {
   systemInfo: SystemInfo,
   coreInfo: CoreSystemInfo,
   cardanoNodeState: ?CardanoNodeState,
-  // isDev: boolean,
-  // isMainnet: boolean,
-  // isStaging: boolean,
-  // isTestnet: boolean,
   isNodeResponding: boolean,
   // isNodeSubscribed: boolean,
   isNodeSyncing: boolean,
@@ -364,12 +335,8 @@ type Props = {
   isSystemTimeCorrect: boolean,
   isSystemTimeIgnored: boolean,
   // isForceCheckingNodeTime: boolean,
-  // latestLocalBlockTimestamp: number,
-  // latestNetworkBlockTimestamp: number,
   localTip: ?TipInfo,
   networkTip: ?TipInfo,
-  // localBlockHeight: number,
-  // networkBlockHeight: number,
   // onForceCheckLocalTimeDifference: Function,
   onOpenStateDirectory: Function,
   onOpenExternalLink: Function,
@@ -391,9 +358,6 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
 
   constructor(props: Props) {
     super(props);
-    // let { localBlockHeight, networkBlockHeight } = props;
-    // localBlockHeight = localBlockHeight || null;
-    // networkBlockHeight = networkBlockHeight || null;
     this.state = {
       isNodeRestarting: false,
     };
@@ -447,7 +411,8 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
         ? intl.formatMessage(messages.statusOn)
         : intl.formatMessage(messages.statusOff);
       className =
-        value || (!value && messageId === 'systemTimeIgnored')
+        (value && messageId !== 'systemTimeIgnored') ||
+        (!value && messageId === 'systemTimeIgnored')
           ? classNames([className, styles.green])
           : classNames([className, styles.red]);
     }
@@ -483,19 +448,11 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
       // isForceCheckingNodeTime,
       localTip,
       networkTip,
-      // localBlockHeight,
-      // networkBlockHeight,
-      // latestLocalBlockTimestamp,
-      // latestNetworkBlockTimestamp,
       onOpenStateDirectory,
       onClose,
       onCopyStateDirectoryPath,
       nodeConnectionError,
       onOpenExternalLink,
-      // isDev,
-      // isTestnet,
-      // isStaging,
-      // isMainnet,
       isCheckingSystemTime,
     } = this.props;
 
@@ -524,49 +481,6 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
     const isNTPServiceReachable = localTimeDifference != null;
     const connectionError = get(nodeConnectionError, 'values', '{}');
     const { message, code } = connectionError;
-
-    // const remainingUnsyncedBlocks = networkBlockHeight - localBlockHeight;
-    // const remainingUnsyncedBlocksClasses = classNames([
-    //   remainingUnsyncedBlocks < 0 ||
-    //   remainingUnsyncedBlocks > UNSYNCED_BLOCKS_ALLOWED
-    //     ? styles.red
-    //     : styles.green,
-    //   styles.layoutData
-    // ]);
-
-    // const latestLocalBlockAge = moment(Date.now()).diff(
-    //   moment(latestLocalBlockTimestamp)
-    // );
-    // const isLocalBlockHeightStalling =
-    //   latestLocalBlockAge > MAX_ALLOWED_STALL_DURATION;
-    // const latestLocalBlockAgeClasses = classNames([
-    //   latestLocalBlockTimestamp > 0 && !isLocalBlockHeightStalling
-    //     ? styles.green
-    //     : styles.red,
-    //   styles.layoutData
-    // ]);
-
-    // const latestNetworkBlockAge = moment(Date.now()).diff(
-    //   moment(latestNetworkBlockTimestamp)
-    // );
-    // const isNetworkBlockHeightStalling =
-    //   latestNetworkBlockAge > MAX_ALLOWED_STALL_DURATION;
-    // const latestNetworkBlockAgeClasses = classNames([
-    //   latestNetworkBlockTimestamp > 0 && !isNetworkBlockHeightStalling
-    //     ? styles.green
-    //     : styles.red,
-    //   styles.layoutData
-    // ]);
-
-    // Cardano Node EKG server is not enabled for the Mainnet and Testnet builds!
-    // const showCardanoNodeEkgLink =
-    //   isMainnet || isTestnet
-    //     ? false
-    //     : getNetworkEkgUrl({
-    //         isDev,
-    //         isStaging,
-    //         isTestnet,
-    //       });
 
     const unknownDiskSpaceSupportUrl = intl.formatMessage(
       messages.unknownDiskSpaceSupportUrl
@@ -723,30 +637,6 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                   {get(localTip, 'slot', '-')}
                 </Fragment>
               )}
-              {/*
-                <div className={styles.layoutRow}>
-                  <div className={styles.layoutHeader}>{intl.formatMessage(messages.remainingUnsyncedBlocks)}:</div>
-                  <div className={remainingUnsyncedBlocksClasses}>
-                    {remainingUnsyncedBlocks >= 0 ? remainingUnsyncedBlocks : '-'}
-                  </div>
-                </div>
-                <div className={styles.layoutRow}>
-                  <div className={styles.layoutHeader}>{intl.formatMessage(messages.latestLocalBlockAge)}:</div>
-                  <div className={latestLocalBlockAgeClasses}>
-                    {latestLocalBlockTimestamp > 0
-                      ? `${latestLocalBlockAge} ms`
-                      : '-'}
-                  </div>
-                </div>
-                <div className={styles.layoutRow}>
-                  <div className={styles.layoutHeader}>{intl.formatMessage(messages.latestNetworkBlockAge)}:</div>
-                  <div className={latestNetworkBlockAgeClasses}>
-                    {latestNetworkBlockTimestamp > 0
-                      ? `${latestNetworkBlockAge} ms`
-                      : '-'}
-                  </div>
-                </div>
-              */}
               <div className={styles.layoutRow}>
                 <div className={styles.layoutHeader}>
                   {intl.formatMessage(messages.localTimeDifference)}:
@@ -803,27 +693,6 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                     : intl.formatMessage(messages.cardanoNodeStatusRestart)}
                 </button>
               )}
-              {/*
-                {showCardanoNodeEkgLink ? (
-                  <div className={styles.layoutRow}>
-                    <div className={styles.layoutHeader}>
-                      {intl.formatMessage(messages.cardanoNodeDiagnostics)}:
-                    </div>
-                    <div className={styles.layoutData}>
-                      <button
-                        className={styles.realTimeStatusBtn}
-                        onClick={() => onOpenExternalLink(showCardanoNodeEkgLink)}
-                      >
-                        {intl.formatMessage(messages.realtimeStatisticsMonitor)}
-                        <SVGInline
-                          svg={externalLinkIcon}
-                          className={styles.externalLinkIcon}
-                        />
-                      </button>
-                    </div>
-                  </div>
-                ) : null}
-              */}
               {getRow(
                 'cardanoNodeState',
                 upperFirst(