+++ source/renderer/app/themes/daedalus/cardano.js
@@ -439,11 +439,6 @@ export const CARDANO_THEME_OUTPUT = {
       'rgba(255, 255, 255, 0.05)',
     '--theme-report-issue-connecting-text-color': '#ffffff',
     '--theme-report-issue-icon-color': '#ffffff',
-    '--theme-report-issue-syncing-background-color': 'rgba(94, 96, 102, 0.05)',
-    '--theme-report-issue-syncing-text-color': '#5e6066',
-    '--theme-report-issue-syncing-download-logs-text-color': '#5e6066',
-    '--theme-report-issue-syncing-download-logs-text-color-connecting':
-      '#ffffff',
   },
   rpAutocomplete: {
     '--rp-autocomplete-bg-color': '#fff',
@@ -546,6 +541,11 @@ export const CARDANO_THEME_OUTPUT = {
     '--rp-radio-label-text-color': '#5e6066',
     '--rp-radio-size': '22px',
   },
+  rpScrollBar: {
+    '--rp-scrollbar-thumb-bg-color:': 'rgba(94, 96, 102, 0.3)',
+    '--rp-scrollbar-thumb-bg-color-active': 'rgba(94, 96, 102, 0.5)',
+    '--rp-scrollbar-thumb-bg-color-hover': 'rgba(94, 96, 102, 0.5)',
+  },
   rpSelect: {
     '--rp-select-arrow-bg-color': 'rgba(94, 96, 102, 0.3)',
     '--rp-select-arrow-bg-color-open': 'rgba(94, 96, 102, 0.7)',
@@ -601,7 +601,6 @@ export const CARDANO_THEME_OUTPUT = {
   },
   scrollbar: {
     '--theme-scrollbar-thumb-background': 'rgba(94, 96, 102, 0.3)',
-    '--theme-scrollbar-thumb-background-hover': 'rgba(94, 96, 102, 0.5)',
   },
   sendConfirmation: {
     '--theme-send-confirmation-dialog-send-values-color': '#ea4c5b',
@@ -640,11 +639,11 @@ export const CARDANO_THEME_OUTPUT = {
     '--theme-sidebar-menu-add-button-background-color-hover': '#272a2e',
     '--theme-sidebar-menu-add-button-text-color': '#ffffff',
     '--theme-sidebar-wallets-scrollbar-background-color':
-      'rgba(255, 255, 255, 0.1)',
+      'rgba(255, 255, 255, 0.3)',
     '--theme-sidebar-wallets-scrollbar-background-color-active':
-      'rgba(255, 255, 255, 0.3)',
+      'rgba(255, 255, 255, 0.5)',
     '--theme-sidebar-wallets-scrollbar-background-color-hover':
-      'rgba(255, 255, 255, 0.3)',
+      'rgba(255, 255, 255, 0.5)',
     '--theme-sidebar-category-networkInfo-background-color': '#2cbb69',
     '--theme-sidebar-category-networkInfo-text-color': '#fff',
   },